# Backend/docs/swagger.yaml
definitions:
...
  controller.AdminServiceResponse:
    properties:
      description:
        type: string
      display_order:
        type: integer
      features:
        items:
          type: string
        type: array
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/domain.Image'
        type: array
      is_published:
        type: boolean
      title:
        type: string
    type: object
  controller.ImagesResponse:
    properties:
      is_cover:
        type: boolean
      url:
        type: string
    type: object
...
  controller.ServiceResponse:
    properties:
      description:
        type: string
      display_order:
        type: integer
      features:
        items:
          type: string
        type: array
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/controller.ImagesResponse'
        type: array
      is_published:
        type: boolean
      title:
        type: string
    type: object
...
  domain.EntityType:
    enum:
    - product
    - blog_news
    - event
    - service
    - testimonial
    type: string
    x-enum-varnames:
    - EntityProduct
    - EntityBlogNews
    - EntityEvent
    - EntityService
    - EntityTestimonial
  domain.ErrorDetails:
    properties:
      code:
        type: string
      details:
        type: string
    type: object
  domain.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/domain.ErrorDetails'
      message:
        default: Error
        type: string
      success:
        default: false
        type: boolean
    type: object
  domain.Image:
    properties:
      caption:
        type: string
      cloudinary_public_id:
        type: string
      created_at:
        type: string
      display_order:
        type: integer
      entity_id:
        type: string
      entity_type:
        $ref: '#/definitions/domain.EntityType'
      id:
        type: string
      is_cover:
        type: boolean
      url:
        type: string
    type: object
...
  domain.Service:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_order:
        type: integer
      features:
        items:
          type: string
        type: array
      id:
        type: string
      is_published:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
    required:
    - title
    type: object
...
  domain.SuccessResponse:
    properties:
      data: {}
      message:
        default: Success
        type: string
      success:
        default: true
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/service:
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Create a new service with details and cover image
      parameters:
      - description: Service title
        in: formData
        name: title
        required: true
        type: string
      - description: Service description
        in: formData
        name: description
        required: true
        type: string
      - collectionFormat: multi
        description: Service features
        in: formData
        items:
          type: string
        name: features
        type: array
      - description: Display order
        in: formData
        name: display_order
        type: integer
      - description: Publish status
        in: formData
        name: is_published
        type: boolean
      - description: Service cover image
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Service'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a new service
      tags:
      - admin service
  /api/admin/services:
    get:
      description: Get all services including unpublished services
      parameters:
      - description: Filter only published services
        in: query
        name: published-only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controller.AdminServiceResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all services (admin)
      tags:
      - admin service
  /api/admin/services/{id}:
    delete:
      description: Delete an existing service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete service
      tags:
      - admin service
    patch:
      consumes:
      - application/json
      - multipart/form-data
      description: Update an existing service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Service title
        in: formData
        name: title
        type: string
      - description: Service description
        in: formData
        name: description
        type: string
      - collectionFormat: multi
        description: Service features
        in: formData
        items:
          type: string
        name: features
        type: array
      - description: Display order
        in: formData
        name: display_order
        type: integer
      - description: Publish status
        in: formData
        name: is_published
        type: boolean
      - description: Service cover image
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Service'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update service
      tags:
      - admin service
...
  /api/services:
    get:
      description: Get all published services available to users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controller.ServiceResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get published services
      tags:
      - service
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"